// ui/src/app/types.d.ts
// types.d.ts

export interface Asset {
    BijihHipotetik: number;
    BijihTerbukti: number;
    BijihTereka: number;
    BijihTerkira: number;
    BijihTertunjuk: number;
    BijihTerukur: number;
    CompetentPerson: string;
    DataAcuan: string;
    IdLogam: string;
    InstansiSumber: string;
    JenisIjin: string;
    JenisKomoditi: string;
    Kabupaten: string;
    Kecamatan: string;
    KelompokLogam: string;
    Keterangan: string;
    LambangUnsur: string;
    Latitude: number;
    LogamHipotetik: number;
    LogamTerbukti: number;
    LogamTereka: number;
    LogamTerkira: number;
    LogamTertunjuk: number;
    LogamTerukur: number;
    LokasiLogam: string;
    Longitude: number;
    NamaObject: string;
    Provinsi: string;
    StatusPenyelidikan: string;
    TahunUpdate: number;
    cid_dataacuan: string;
    filename_acuan: string;
    randGnum: string;
    sk_agn: string;
    sk_bgn: string;
    status_enc: number;
    status_validasi: number;
    username: string;
  }
  